% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netgen.R
\name{netgen}
\alias{netgen}
\title{Generate Simulated Nonlinear Network Data}
\usage{
netgen(
  n,
  mean_cent,
  sd_cent,
  mean_peri_non,
  mean_peri_lin,
  sd_peri_non,
  sd_peri_lin,
  mean_target,
  sd_target,
  beta_peri_lin,
  beta_cent_lin,
  beta_lin,
  beta_non,
  beta_cent_non,
  beta_peri_non,
  inter_cent_lin,
  beta_con,
  func,
  con_func
)
}
\arguments{
\item{n}{Integer. Number of observations to generate.}

\item{mean_cent}{Mean of the central variable \code{A}.}

\item{sd_cent}{Standard deviation of the central variable \code{A}.}

\item{mean_peri_non}{Numeric vector. Means of the peripheral nonlinear variables.}

\item{mean_peri_lin}{Numeric vector. Means of the peripheral linear variables.}

\item{sd_peri_non}{Numeric vector. Standard deviations of the peripheral nonlinear variables.}

\item{sd_peri_lin}{Numeric vector. Standard deviations of the peripheral linear variables.}

\item{mean_target}{Mean of the random error term for the target variable \code{C}.}

\item{sd_target}{Standard deviation of the random error term for the target variable \code{C}.}

\item{beta_peri_lin}{Numeric vector. Regression coefficients for effects of the peripheral linear variables on \code{C}.}

\item{beta_cent_lin}{Numeric vector. Regression coefficients for effects of \code{A} on the peripheral linear variables.}

\item{beta_lin}{Coefficient for the linear effect of \code{A} on \code{C}.}

\item{beta_non}{Coefficient for the nonlinear effect of \code{A} on \code{C}.}

\item{beta_cent_non}{Numeric vector. Regression coefficients for effects of \code{A} on the peripheral nonlinear variables.}

\item{beta_peri_non}{Numeric vector. Regression coefficients for effects of the peripheral nonlinear variables on \code{C}.}

\item{inter_cent_lin}{Integer vector. Indices identifying peripheral variables involved in interactions with \code{A}.}

\item{beta_con}{Regression coefficient for the confounding effect.}

\item{func}{Character. Functional form relating \code{A} to \code{C}.}

\item{con_func}{Character. Functional form relating \code{A} to the nonlinear peripheral variables.}
}
\value{
A data frame containing \code{A}, \code{C}, linear peripheral variables
\code{B1}, \code{B2}, ..., and nonlinear peripheral variables \code{D1}, \code{D2}, ....
}
\description{
Simulates a dataset containing a central variable, peripheral variables,
and a target variable under different functional relationships.
}
\examples{
set.seed(123)

dat <- netgen(
  n = 100,
  mean_cent = 0,
  sd_cent = 1,
  mean_peri_non = rep(0, 4),
  mean_peri_lin = rep(0, 3),
  sd_peri_non = rep(1, 4),
  sd_peri_lin = rep(1, 3),
  mean_target = 0,
  sd_target = 1,
  beta_peri_lin = c(0.3, 0.4, 0.5),
  beta_cent_lin = rep(0.5, 3),
  beta_lin = 0.5,
  beta_non = 0.75,
  beta_cent_non = c(0.5, 0.6, 0.7, 0.8),
  beta_peri_non = c(0.3, 0.4, 0.5, 0.6),
  inter_cent_lin = c(1, 2, 3),
  beta_con = 0,
  func = "quad",
  con_func = "quad"
)

head(dat)

}
