% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcor_res.R
\name{dcor_res}
\alias{dcor_res}
\title{Distance Correlation Test After Residualization}
\usage{
dcor_res(dat, A, C, B_variables, nperm)
}
\arguments{
\item{dat}{A data frame}

\item{A}{Character string specifying the name of the first variable to be used in the distance correlation.}

\item{C}{Character string specifying the name of the second variable to be used in the distance correlation.}

\item{B_variables}{Character vector containing the names of the covariates
used as predictors when residualizing \code{A} and \code{C}.}

\item{nperm}{Integer. Number of permutations used in the distance
correlation test.}
}
\value{
A numeric value giving the permutation p-value from
\code{energy::dcor.test()}.
}
\description{
Residualizes two random variables with respect to a set of covariates using
generalized additive models, then tests the association between the
residualized variables using a permutation-based distance correlation test.
}
